backend/requirements.txt
# नेपाल Hostel Finder — Backend Dependencies
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
Pillow>=10.0,<11.0
python-decouple>=3.8,<4.0
psycopg2-binary>=2.9,<3.0
google-auth>=2.29,<3.0
python-dotenv>=1.0,<2.0
